Update from upstream repo facebook/create-react-app@master

+++ packages/react-scripts/scripts/utils/verifyTypeScriptSetup.js
@@ -101,6 +101,7 @@ function verifyTypeScriptSetup() {
     forceConsistentCasingInFileNames: { suggested: true },
 
     // These values are required and cannot be changed by the user
+    // Keep this in sync with the webpack config
     module: {
       parsedValue: ts.ModuleKind.ESNext,
       value: 'esnext',
@@ -119,6 +120,13 @@ function verifyTypeScriptSetup() {
       value: 'preserve',
       reason: 'JSX is compiled by Babel',
     },
+    // We do not support absolute imports, though this may come as a future
+    // enhancement
+    baseUrl: {
+      value: undefined,
+      reason: 'absolute imports are not supported (yet)',
+    },
+    paths: { value: undefined, reason: 'aliased imports are not supported' },
   };
 
   const formatDiagnosticHost = {
@@ -183,12 +191,13 @@ function verifyTypeScriptSetup() {
     const { parsedValue, value, suggested, reason } = compilerOptions[option];
 
     const valueToCheck = parsedValue === undefined ? value : parsedValue;
+    const coloredOption = chalk.cyan('compilerOptions.' + option);
 
     if (suggested != null) {
       if (parsedCompilerOptions[option] === undefined) {
         appTsConfig.compilerOptions[option] = suggested;
         messages.push(
-          `${chalk.cyan('compilerOptions.' + option)} to be ${chalk.bold(
+          `${coloredOption} to be ${chalk.bold(
             'suggested'
           )} value: ${chalk.cyan.bold(suggested)} (this can be changed)`
         );
@@ -196,9 +205,9 @@ function verifyTypeScriptSetup() {
     } else if (parsedCompilerOptions[option] !== valueToCheck) {
       appTsConfig.compilerOptions[option] = value;
       messages.push(
-        `${chalk.cyan('compilerOptions.' + option)} ${chalk.bold(
-          'must'
-        )} be ${chalk.cyan.bold(value)}` +
+        `${coloredOption} ${chalk.bold(
+          valueToCheck == null ? 'must not' : 'must'
+        )} be ${valueToCheck == null ? 'set' : chalk.cyan.bold(value)}` +
           (reason != null ? ` (${reason})` : '')
       );
     }
@@ -238,21 +247,13 @@ function verifyTypeScriptSetup() {
     writeJson(paths.appTsConfig, appTsConfig);
   }
 
-  // Copy type declarations associated with this version of `react-scripts`
-  const declaredTypes = path.resolve(
-    __dirname,
-    '..',
-    '..',
-    'config',
-    'react-app.d.ts'
-  );
-  const declaredTypesContent = fs
-    .readFileSync(declaredTypes, 'utf8')
-    .replace(/\/\/ @remove-file-on-eject\r?\n/, '');
-  fs.writeFileSync(
-    path.resolve(paths.appSrc, 'react-app.d.ts'),
-    declaredTypesContent
-  );
+  // Reference `react-scripts` types
+  if (!fs.existsSync(paths.appTypeDeclarations)) {
+    fs.writeFileSync(
+      paths.appTypeDeclarations,
+      `/// <reference types="react-scripts" />${os.EOL}`
+    );
+  }
 }
 
 module.exports = verifyTypeScriptSetup;
